chore(ui-ux): cleanup add/remove liquidity v1
