+++ mobile-app/app/screens/AppNavigator/screens/Dex/DexNavigator.tsx
@@ -19,15 +19,10 @@ import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import { Platform, StyleProp, ViewStyle } from 'react-native'
 import { ThemedTextV2 } from '@components/themed'
 import { tailwind } from '@tailwind'
-import { AddLiquidityScreenV2 } from './DexAddLiquidityV2'
 import { ConfirmAddLiquidityScreen } from './DexConfirmAddLiquidity'
 import { RemoveLiquidityScreen } from './DexRemoveLiquidity'
-import { RemoveLiquidityScreenV2 } from './DexRemoveLiquidityV2'
 import { RemoveLiquidityConfirmScreen } from './DexConfirmRemoveLiquidity'
-import { RemoveLiquidityConfirmScreenV2 } from './DexConfirmRemoveLiquidityV2'
-import { useFeatureFlagContext } from '@contexts/FeatureFlagContext'
 import { AddLiquidityScreen } from './DexAddLiquidity'
-import { ConfirmAddLiquidityScreenV2 } from './DexConfirmAddLiquidityV2'
 
 export interface DexParamList {
   DexScreen: undefined
@@ -105,7 +100,6 @@ const DexStack = createStackNavigator<DexParamList>()
 export function DexNavigator (): JSX.Element {
   const navigation = useNavigation<NavigationProp<DexParamList>>()
   const headerContainerTestId = 'dex_header_container'
-  const { isFeatureAvailable } = useFeatureFlagContext()
   const screenOptions = useNavigatorScreenOptions()
   const goToNetworkSelect = (): void => {
     navigation.navigate('NetworkSelectionScreen')
@@ -160,7 +154,7 @@ export function DexNavigator (): JSX.Element {
       />
 
       <DexStack.Screen
-        component={isFeatureAvailable('add_liquidity_v2') ? AddLiquidityScreenV2 : AddLiquidityScreen}
+        component={AddLiquidityScreen}
         name='AddLiquidity'
         options={{
           ...screenOptions,
@@ -172,7 +166,7 @@ export function DexNavigator (): JSX.Element {
       />
 
       <DexStack.Screen
-        component={isFeatureAvailable('add_liquidity_v2') ? ConfirmAddLiquidityScreenV2 : ConfirmAddLiquidityScreen}
+        component={ConfirmAddLiquidityScreen}
         name='ConfirmAddLiquidity'
         options={{
           ...screenOptions,
@@ -184,7 +178,7 @@ export function DexNavigator (): JSX.Element {
       />
 
       <DexStack.Screen
-        component={isFeatureAvailable('remove_liquidity_v2') ? RemoveLiquidityScreenV2 : RemoveLiquidityScreen}
+        component={RemoveLiquidityScreen}
         name='RemoveLiquidity'
         options={{
           ...screenOptions,
@@ -196,7 +190,7 @@ export function DexNavigator (): JSX.Element {
       />
 
       <DexStack.Screen
-        component={isFeatureAvailable('remove_liquidity_v2') ? RemoveLiquidityConfirmScreenV2 : RemoveLiquidityConfirmScreen}
+        component={RemoveLiquidityConfirmScreen}
         name='RemoveLiquidityConfirmScreen'
         options={{
           ...screenOptions,
